+++ torchpipe/csrc/backend/include/OpenvinoMat.hpp
@@ -16,41 +16,18 @@
 
 #include "Backend.hpp"
 #include "params.hpp"
-#include "prepost.hpp"
+
 namespace ipipe {
-struct ModelInstances;
+
+namespace o_v {
+class ModelInstances;
+}
 
 class OpenvinoMat : public Backend {
  public:
-  /**
-   * @brief
-   * @param postprocessor
-   * 自定义后处理。自定义网络输出c++批后处理;默认操作是拆分batch维度；
-   * 需要实现为 @ref PostProcessor  的子类并注册
-   * @param OpenvinoMat::backend 模型转换器。@ref _engine
-   * 不存在时，会调用此参数指定的转换器，默认为 @ref Onnx2TensorrtConverter.
-   * @param instance_num 实例数目；如果tensorrt
-   * engine的profile数目不足以建立足够的实例数，将反序列化多个engine.
-   * @param _engine
-   类型为`std::shared_ptr<CudaEngineWithRuntime>`, 位于shared_config 中的
-   tensorrt 引擎. 如果此tensorrt engine
-   已经没有profile供下一个实例使用时，会删除此键值。 如果不存在， 将调用
-   @ref OpenvinoMat::backend 指定的转换器产生。
-   @param _independent_thread_index 实例序号。 默认为0.
-   */
   virtual bool init(const std::unordered_map<std::string, std::string> &,
                     dict shared_config) override;
 
-  /**
-   * @brief
-   * @param TASK_DATA_KEY
-   * 输入数据。网络为单输入输出时，类型为at::Tensor/torch.Tensor,
-   * 网络为多输入输出时， 类型为vector/List. 字典序排列.
-   * @param[out] TASK_RESULT_KEY 输出类型参照输入类型。 gpu上连续, @ref
-   * postprocessor 可修改此类型。网络多输出时按照名称的字典序排列。
-   * @ref postprocessor 可自定义输出。需要确保给 TASK_RESULT_KEY 赋值。
-   *
-   */
   void forward(const std::vector<dict> &) override;
 
   /**
@@ -62,10 +39,7 @@ class OpenvinoMat : public Backend {
    * @brief 此实例中模型的最大小batch
    *
    */
-  virtual uint32_t min() const {
-    return 1;
-    ;
-  };
+  virtual uint32_t min() const { return 1; };
   ~OpenvinoMat(){
 
   };
@@ -74,21 +48,8 @@ class OpenvinoMat : public Backend {
   std::unique_ptr<Params> params_;
 
   std::unique_ptr<Backend> backend_;
-  std::vector<std::vector<int>> mins_;
-  std::vector<std::vector<int>> maxs_;
-  std::shared_ptr<ModelInstances> instances_;
 
-  int max_{1};
-  int min_{1};
-
-  std::vector<void *> binding_;
-
-  std::vector<bool> change_shape_;
-
-  int profile_index_{0};
-
-  std::vector<uint32_t> new_location_inputs_;
-  std::vector<uint32_t> new_location_ouputs_;
+  std::shared_ptr<ipipe::o_v::ModelInstances> instances_;
 
   int independent_thread_index_{0};
 };