+++ torchpipe/csrc/backend/src/OpenvinoMat.cpp
@@ -31,23 +31,35 @@
 #include "base_logging.hpp"
 #include "exception.hpp"
 #include "OpenvinoMat.hpp"
-#include <openvino/openvino.hpp>
 
+#include "cnn.hpp"
+#include <opencv2/core.hpp>
 // https://github.com/NVIDIA/Torch-TensorRT/blob/3a98a8b198a071e622c43283caea7416fe8a8a1a/core/runtime/register_trt_op.cpp
 
 // #define USE_OUT_MEM
 namespace ipipe {
 
+namespace o_v {
 struct ModelInstances {
-  ov::CompiledModel compiled_model;
-  std::vector<ov::InferRequest> ireqs;
+  ModelInstances(std::string model_path, ov::Core& core, int instance_num) : core_(core) {
+    CnnConfig config(model_path);
+    config.m_core = core;
+    config.instance_num = instance_num;
+    config.m_deviceName = "CPU";
+    model = std::make_unique<VectorCNN>(config);
+  }
+
+  std::unique_ptr<VectorCNN> model;
+
+ private:
+  ov::Core& core_;
 };
+}  // namespace o_v
+using ipipe::o_v::ModelInstances;
 
 bool OpenvinoMat::init(const std::unordered_map<std::string, std::string>& config_param,
                        dict dict_config) {
-  params_ = std::unique_ptr<Params>(new Params({{"postprocessor", "split"},
-                                                {"preprocessor", ""},
-                                                {"_independent_thread_index", "0"},
+  params_ = std::unique_ptr<Params>(new Params({{"_independent_thread_index", "0"},
                                                 {"OpenvinoMat::backend", ""},
                                                 {"instance_num", "1"},
                                                 {"precision", "fp32"}},
@@ -78,39 +90,33 @@ bool OpenvinoMat::init(const std::unordered_map<std::string, std::string>& confi
   std::string model = params_->at("model");
   // get engine
   auto& config = *dict_config;
-  ov::Core core;
-  // core.set_property("CPU", ov::streams::num(ov::streams::AUTO));
 
-  instances_ = std::make_shared<ModelInstances>();
+  // core.set_property("CPU", ov::streams::num(ov::streams::AUTO));
+  static ov::Core core;
+  IPIPE_ASSERT(instance_num == 1);
+
+  // input_shape_ = instances_->.ov_model.input().get_shape();
+  // input_shape_[ov::layout::batch_idx({"NHWC"})] = 1;
 
   if (config.count("_engine") == 0) {
-    instances_->compiled_model = core.compile_model(
-        model, "CPU", ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT),
-        // ov::streams::num(instance_num), ov::inference_num_threads(instance_num),
-        ov::hint::num_requests(instance_num), ov::hint::allow_auto_batching(false));
+    instances_ = std::make_shared<ModelInstances>(model, core, instance_num);
+    // instances_->compiled_model = core.compile_model(
+    //     model, "CPU", ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT),
+    //     // ov::streams::num(instance_num), ov::inference_num_threads(instance_num),
+    //     ov::hint::num_requests(instance_num), ov::hint::allow_auto_batching(false));
 
-    auto nthreads = instances_->compiled_model.get_property(ov::inference_num_threads);
-    // IPIPE_ASSERT(nthreads == instance_num);
+    // auto nthreads = instances_->compiled_model.get_property(ov::inference_num_threads);
+    // // IPIPE_ASSERT(nthreads == instance_num);
 
-    auto nireq = instances_->compiled_model.get_property(ov::optimal_number_of_infer_requests);
+    // auto nireq = instances_->compiled_model.get_property(ov::optimal_number_of_infer_requests);
 
-    if (nireq != instance_num) {
-      SPDLOG_WARN(
-          "optimal number of requests does not match the `instance_num`. It is "
-          "recommended to set the number of instances equal to the recommended optimal number of "
-          "requests");
-    }
-
-    // /* Take information about all topology inputs */
-    // auto inputs = model->inputs();
-    // /* Take information about all topology outputs */
-    // auto outputs = model->outputs();
-    // ov::Shape static_shape;
-    // ov::PartialShape partial_shape =
-    //     node->output(0).get_partial_shape();  // get zero output partial shape
-    // if (!partial_shape.is_dynamic() /* or partial_shape.is_static() */) {
-    //   static_shape = partial_shape.get_shape();
+    // if (nireq != instance_num) {
+    //   SPDLOG_WARN(
+    //       "optimal number of requests does not match the `instance_num`. It is "
+    //       "recommended to set the number of instances equal to the recommended optimal number of
+    //       " "requests");
     // }
+
     config["_engine"] = instances_;
   } else {
     instances_ = any_cast<std::shared_ptr<ModelInstances>>(config.at("_engine"));
@@ -120,10 +126,32 @@ bool OpenvinoMat::init(const std::unordered_map<std::string, std::string>& confi
   assert(config.count("_engine") != 0);
 
   std::string precision = params_->at("precision");
+  // https://docs.openvino.ai/2023.0/groupov_dev_api_system_conf.html#doxid-group-ov-dev-api-system-conf-1gad1a071adcef91309ca90878afd83f4fe
+  // if (precision == "bf16") {
+  //   IPIPE_ASSERT(ov::with_cpu_x86_bfloat16() || ov::with_cpu_x86_avx512_core_amx_bf16());
+  // } else if (precision == "int8") {
+  //   IPIPE_ASSERT(ov::with_cpu_x86_avx512_core_vnni() || ov::with_cpu_x86_avx512_core_amx_int8());
+  // }
+  // else if (precision == "fp32"){
+  //   IPIPE_ASSERT(ov::with_cpu_x86_avx512f()||ov::with_cpu_x86_avx512_core_amx());
+
+  IPIPE_ASSERT(precision == "fp32");  // todo: support bf16, int8
 
   return true;
 }
-void OpenvinoMat::forward(const std::vector<dict>&) {}
+void OpenvinoMat::forward(const std::vector<dict>& input_dicts) {
+  cv::Mat input = any_cast<cv::Mat>(input_dicts[0]->at(TASK_DATA_KEY));
+  if (input.isContinuous()) {
+    input = input.clone();
+  }
+  IPIPE_ASSERT(input.elemSize1() == 1 || input.elemSize1() == 4);
+  cv::Mat output;
+  instances_->model->Compute(input, &output);
+
+  (*input_dicts[0])[TASK_RESULT_KEY] = output;
+
+  // https://docs.openvino.ai/2022.3/openvino_docs_OV_UG_Preprocessing_Overview.html#doxid-openvino-docs-o-v-u-g-preprocessing-overview
+}
 
 IPIPE_REGISTER(Backend, OpenvinoMat, "OpenvinoMat");